tracking

+++ app/models/beatmap.go
@@ -4,6 +4,6 @@ import "github.com/jinzhu/gorm"
 
 type Beatmap struct {
 	gorm.Model
-	BeatmapID    int64
-	BeatmapSetID int64
+	BeatmapID    int
+	BeatmapSetID int
 }
